monolithic Database classes changed to individual handlers per collection
