+++ db/task_handler.py
@@ -1,0 +1,36 @@
+from .base_handler import BaseHandler
+
+
+class TaskHandler(BaseHandler):
+    """Handler for task history and updates"""
+    
+    def __init__(self, uri, db_name):
+        super().__init__(uri, db_name)
+        self.task_updates_collection = self.db['task_updates']
+    
+    def get_task_history(self, hypothesis_id):
+        """Get task history for hypothesis"""
+        task_history = list(self.task_updates_collection.find({"hypothesis_id": hypothesis_id}))
+        
+        for update in task_history:
+            update["_id"] = str(update["_id"])
+        return task_history
+    
+    def get_latest_task_state(self, hypothesis_id):
+        """Get latest task state for hypothesis"""
+        task_history = list(self.task_updates_collection.find({"hypothesis_id": hypothesis_id}).sort("timestamp", -1).limit(1))
+        if task_history:
+            return task_history[0]
+        return None
+
+    def save_task_history(self, hypothesis_id, task_history):
+        """Save complete task history to DB"""
+        # Delete existing history first
+        self.task_updates_collection.delete_many({"hypothesis_id": hypothesis_id})
+        
+        # Insert new history as a batch
+        if task_history:
+            self.task_updates_collection.insert_many([
+                {**update, "hypothesis_id": hypothesis_id}
+                for update in task_history
+            ])
